Add publish subcommand
