+++ README.md
@@ -848,7 +848,7 @@ Sync documentation:
   commands:
     - for file in $(find -type f -name '*.md'); do
         echo "> Sync $file";
-        mark -u $MARK_USER -p $MARK_PASS -b $MARK_URL -f $file || exit 1;
+        mark publish -u $MARK_USER -p $MARK_PASS -b $MARK_URL -f $file || exit 1;
         echo;
       done
 ```
@@ -859,7 +859,7 @@ repository with documentation will be cloned to. The following command finds all
 ```bash
 for file in $(find -type f -name '*.md'); do
     echo "> Sync $file";
-    mark -u $MARK_USER -p $MARK_PASS -b $MARK_URL -f $file || exit 1;
+    mark publish -u $MARK_USER -p $MARK_PASS -b $MARK_URL -f $file || exit 1;
     echo;
 done
 ```
@@ -880,13 +880,13 @@ only:
 Rather than running `mark` multiple times, or looping through a list of files from `find`, you can use file globbing (i.e. wildcard patterns) to match files in subdirectories. For example:
 
 ```bash
-mark -f "helpful_cmds/*.md"
+mark publish -f "helpful_cmds/*.md"
 ```
 
 You can also use `**` to get all files recursively.
 
 ```bash
-mark -f "**/docs/*.md"
+mark publish -f "**/docs/*.md"
 ```
 
 ### Linting markdown
